Move errors into own file.
